+++ R/output.R
@@ -286,8 +286,12 @@ process_file = function(text, output) {
   # was not appropriate for non-interactive mode, and I don't want to argue)
   progress = opts_knit$get('progress') && !is_R_CMD_check()
   if (progress) {
-    pb = txtProgressBar(0, n, char = '.', style = 3)
-    on.exit(close(pb), add = TRUE)
+    labels = unlist(lapply(groups, function(g) {
+      if (is.list(g$params)) g[[c('params', 'label')]] else ''
+    }))
+    pb_fun = getOption('knitr.progress.fun', txt_pb)
+    pb = if (is.function(pb_fun)) pb_fun(n, labels)
+    on.exit(if (!is.null(pb)) pb$done(), add = TRUE)
   }
   wd = getwd()
   for (i in 1:n) {
@@ -299,11 +303,7 @@ process_file = function(text, output) {
       }
       break  # must have called knit_exit(), so exit early
     }
-    if (progress) {
-      setTxtProgressBar(pb, i)
-      if (!tangle) cat('\n')  # under tangle mode, only show one progress bar
-      flush.console()
-    }
+    if (progress && !is.null(pb)) pb$update(i)
     group = groups[[i]]
     res[i] = withCallingHandlers(
       if (tangle) process_tangle(group) else process_group(group),